+++ WalletWasabi/Crypto/OwnershipProof.cs
@@ -7,109 +7,6 @@ using NBitcoin.Crypto;
 
 namespace WalletWasabi.Crypto
 {
-	public class OwnershipIdentifier : IBitcoinSerializable
-	{
-		public const int OwnershipIdLength = 32;
-		private byte[] _bytes;
-
-		public OwnershipIdentifier(byte[] bytes)
-		{
-			if (bytes.Length != OwnershipIdLength)
-			{
-				throw new ArgumentException($"Ownership identifier must be {OwnershipIdLength} bytes long.");
-			}
-
-			_bytes = bytes.ToArray();
-		}
-
-		public OwnershipIdentifier()
-			: this(new byte[OwnershipIdLength])
-		{
-		}
-
-		public void ReadWrite(BitcoinStream bitcoinStream)
-		{
-			bitcoinStream.ReadWrite(ref _bytes);
-		}
-	}
-
-	public record CoinJoinInputCommitmentData
-	{
-		private byte[] _coordinatorIdentifier;
-		private byte[] _roundIdentifier;
-
-		public CoinJoinInputCommitmentData(string coordinatorIdentifier, uint256 roundIdentifier)
-			: this(Encoding.ASCII.GetBytes(coordinatorIdentifier), roundIdentifier.ToBytes())
-		{
-		}
-
-		public CoinJoinInputCommitmentData(byte[] coordinatorIdentifier, byte[] roundIdentifier)
-		{
-			_coordinatorIdentifier = coordinatorIdentifier;
-			_roundIdentifier = roundIdentifier;
-		}
-
-		public byte[] ToBytes() =>
-			BitConverter.GetBytes(_coordinatorIdentifier.Length)
-				.Concat(_coordinatorIdentifier)
-				.Concat(_roundIdentifier)
-				.ToArray();
-	}
-
-	[Flags]
-	public enum ProofBodyFlags : byte
-	{
-		UserConfirmation = 1
-	}
-
-	public class ProofBody : IBitcoinSerializable
-	{
-		private static readonly byte[] VersionMagic = { 0x53, 0x4c, 0x00, 0x19 };
-		private List<OwnershipIdentifier> _ownershipIdentifiers = new();
-		private byte _flags;
-
-		public ProofBody()
-		{
-		}
-
-		public ProofBody(ProofBodyFlags flags, params OwnershipIdentifier[] ownershipIdentifiers)
-		{
-			Flags = flags;
-			_ownershipIdentifiers = ownershipIdentifiers.ToList();
-		}
-
-		public IEnumerable<OwnershipIdentifier> OwnershipIdentifiers => _ownershipIdentifiers;
-
-		public ProofBodyFlags Flags
-		{
-			get => (ProofBodyFlags)_flags;
-			set => _flags = (byte)value;
-		}
-
-		public void ReadWrite(BitcoinStream bitcoinStream)
-		{
-			var versionMagic = VersionMagic.ToArray();
-			bitcoinStream.ReadWrite(ref versionMagic);
-
-			if (!bitcoinStream.Serializing)
-			{
-				if (!VersionMagic.SequenceEqual(versionMagic))
-				{
-					throw new FormatException("Invalid version magic.");
-				}
-			}
-
-			bitcoinStream.ReadWrite(ref _flags);
-			bitcoinStream.ReadWrite(ref _ownershipIdentifiers);
-		}
-
-		public uint256 SignatureHash(Script scriptPubKey, byte[] commitmentData) =>
-			new(Hashes.SHA256(this.ToBytes().Concat(ProofFooter(scriptPubKey, commitmentData)).ToArray()));
-
-		private static IEnumerable<byte> ProofFooter(Script scriptPubKey, byte[] commitmentData) =>
-			scriptPubKey.ToBytes().Concat(commitmentData);
-	}
-
 	public class OwnershipProof : IBitcoinSerializable
 	{
 		private ProofBody _proofBody;
