+++ WalletWasabi/Crypto/ProofBodyFlags.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace WalletWasabi.Crypto
+{
+	[Flags]
+	public enum ProofBodyFlags : byte
+	{
+		UserConfirmation = 1
+	}
+}
